requirements.txt
asgiref==3.7.2
autopep8==2.0.4
backports.zoneinfo==0.2.1
Django==4.2.9
django-prometheus==2.3.1
django-rest-framework==0.1.0
django-silk==5.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
Faker==22.6.0
gprof2dot==2022.7.29
gunicorn==21.2.0
inflection==0.5.1
mysqlclient==2.2.1
packaging==23.2
prometheus-client==0.19.0
pycodestyle==2.11.1
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.4
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
uritemplate==4.1.1